+++ changelog.md
@@ -1,6 +1,12 @@
 Codion Change Log
 ==================
 
+## 0.17.21-SNAPSHOT
+### is.codion.swing.framework.ui
+- EntityPanel bug fixed, table refresh control was always enabled, instead of only when the panel was active.
+- EntityPanel.createControls() added, related refactoring.
+- EntityPanel.createControlsComponent() renamed createControlComponent().
+
 ## 0.17.20
 ### is.codion.common.model
 - DefaultColumnConditionModel.accept() no longer returns true in case the condition model is disabled.